refactor : Party 디렉토리 구조 변경
ui -> join, room 별도로 관리
