+++ feature/party/src/main/java/online/partyrun/partyrunapplication/feature/party/room/PartyRoomViewModel.kt
@@ -1,4 +1,4 @@
-package online.partyrun.partyrunapplication.feature.party.ui
+package online.partyrun.partyrunapplication.feature.party.room
 
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
@@ -8,7 +8,6 @@ import kotlinx.coroutines.CompletableDeferred
 import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.StateFlow
 import kotlinx.coroutines.flow.asStateFlow
-import kotlinx.coroutines.flow.update
 import kotlinx.coroutines.launch
 import online.partyrun.partyrunapplication.core.domain.party.ConnectPartyEventSourceUseCase
 import online.partyrun.partyrunapplication.core.domain.party.CreatePartyEventSourceListenerUseCase
